requirements.txt
pyqt5
opencv-python
scikit-learn
numpy
pillow